refactor: delete dead re-export roots and merge tiny files (12 deleted)
lib/ cleanup:
- Delete 4 dead re-export roots: decisions.py, phases.py, work_events.py,
  raw_payload_sampling.py (0-2 importers each)
- Merge 4 tiny model files into extraction modules:
  decision_models→decision_extraction, phase_models→phase_extraction,
  work_event_models→work_event_extraction, action_event_models→action_events

schemas/ cleanup:
- Merge generation_annotations.py (46 lines) into generation_support.py
- Delete generation_analysis.py (29 lines, importers rewired to sources)
- Merge generation_emission.py (27 lines) into generation_workflow.py
- Merge generation_provider_bundle_failures.py (31 lines) into
  generation_provider_bundle.py

Kept: branch_type.py, filter_types.py (circular import risk if merged),
field_stats.py, observation.py (15+ importers, legitimate hubs).

All 4,320 tests pass.
