--- polylogue/lib/decisions.py
@@ -1,6 +1,0 @@
-"""Public decision semantic surface."""
-
-from polylogue.lib.decision_extraction import extract_decisions
-from polylogue.lib.decision_models import Decision
-
-__all__ = ["Decision", "extract_decisions"]
