--- polylogue/lib/phases.py
@@ -1,6 +1,0 @@
-"""Public phase semantic surface."""
-
-from polylogue.lib.phase_extraction import extract_phases
-from polylogue.lib.phase_models import SessionPhase
-
-__all__ = ["SessionPhase", "extract_phases"]
