+++ polylogue/lib/work_event_extraction.py
@@ -2,6 +2,9 @@
 
 from __future__ import annotations
 
+from dataclasses import dataclass
+from datetime import date, datetime
+from enum import Enum
 from typing import TYPE_CHECKING
 
 from polylogue.lib.phase_extraction import extract_phases
@@ -10,7 +13,89 @@ from polylogue.lib.semantic_facts import (
     MessageSemanticFacts,
     build_conversation_semantic_facts,
 )
-from polylogue.lib.work_event_models import WorkEvent, WorkEventKind
+
+
+class WorkEventKind(str, Enum):
+    """Classification of work performed in a message range."""
+
+    PLANNING = "planning"
+    IMPLEMENTATION = "implementation"
+    DEBUGGING = "debugging"
+    REVIEW = "review"
+    TESTING = "testing"
+    RESEARCH = "research"
+    CONFIGURATION = "configuration"
+    DOCUMENTATION = "documentation"
+    REFACTORING = "refactoring"
+    DATA_ANALYSIS = "data_analysis"
+    CONVERSATION = "conversation"
+
+
+@dataclass(frozen=True)
+class WorkEvent:
+    """A classified segment of work within a conversation."""
+
+    kind: WorkEventKind
+    start_index: int
+    end_index: int
+    confidence: float
+    evidence: tuple[str, ...]
+    file_paths: tuple[str, ...]
+    tools_used: tuple[str, ...]
+    summary: str
+    start_time: datetime | None = None
+    end_time: datetime | None = None
+    canonical_session_date: date | None = None
+    duration_ms: int = 0
+
+    def to_dict(self) -> dict[str, object]:
+        return {
+            "kind": self.kind.value,
+            "start_index": self.start_index,
+            "end_index": self.end_index,
+            "start_time": self.start_time.isoformat() if self.start_time else None,
+            "end_time": self.end_time.isoformat() if self.end_time else None,
+            "canonical_session_date": (
+                self.canonical_session_date.isoformat()
+                if self.canonical_session_date
+                else None
+            ),
+            "duration_ms": self.duration_ms,
+            "confidence": self.confidence,
+            "evidence": list(self.evidence),
+            "file_paths": list(self.file_paths),
+            "tools_used": list(self.tools_used),
+            "summary": self.summary,
+        }
+
+    @classmethod
+    def from_dict(cls, payload: dict[str, object]) -> WorkEvent:
+        return cls(
+            kind=WorkEventKind(str(payload["kind"])),
+            start_index=int(payload.get("start_index", 0) or 0),
+            end_index=int(payload.get("end_index", 0) or 0),
+            start_time=(
+                datetime.fromisoformat(str(payload["start_time"]))
+                if payload.get("start_time")
+                else None
+            ),
+            end_time=(
+                datetime.fromisoformat(str(payload["end_time"]))
+                if payload.get("end_time")
+                else None
+            ),
+            canonical_session_date=(
+                date.fromisoformat(str(payload["canonical_session_date"]))
+                if payload.get("canonical_session_date")
+                else None
+            ),
+            duration_ms=int(payload.get("duration_ms", 0) or 0),
+            confidence=float(payload.get("confidence", 0.0) or 0.0),
+            evidence=tuple(str(item) for item in payload.get("evidence", []) or []),
+            file_paths=tuple(str(item) for item in payload.get("file_paths", []) or []),
+            tools_used=tuple(str(item) for item in payload.get("tools_used", []) or []),
+            summary=str(payload.get("summary", "") or ""),
+        )
 
 if TYPE_CHECKING:
     from polylogue.lib.models import Conversation
@@ -228,4 +313,4 @@ def extract_work_events(
     return _merge_adjacent(events)
 
 
-__all__ = ["extract_work_events"]
+__all__ = ["WorkEvent", "WorkEventKind", "extract_work_events"]
