--- polylogue/lib/work_events.py
@@ -1,6 +1,0 @@
-"""Public work-event semantic surface."""
-
-from polylogue.lib.work_event_extraction import extract_work_events
-from polylogue.lib.work_event_models import WorkEvent, WorkEventKind
-
-__all__ = ["WorkEvent", "WorkEventKind", "extract_work_events"]
