--- polylogue/schemas/generation_analysis.py
@@ -1,29 +1,0 @@
-"""Small public root for schema-generation cluster analysis families."""
-
-from __future__ import annotations
-
-from polylogue.schemas.generation_cluster_collection import (
-    _collect_cluster_accumulators,
-    collect_cluster_analysis,
-)
-from polylogue.schemas.generation_cluster_support import (
-    _artifact_priority,
-    _cluster_profile_tokens,
-    _cluster_reservoir_size,
-    _cluster_sort_key,
-    _merge_representative_paths,
-    _parse_observed_at,
-)
-from polylogue.schemas.sampling import iter_schema_units
-
-__all__ = [
-    "collect_cluster_analysis",
-    "_artifact_priority",
-    "_cluster_profile_tokens",
-    "_cluster_reservoir_size",
-    "_cluster_sort_key",
-    "_collect_cluster_accumulators",
-    "_merge_representative_paths",
-    "_parse_observed_at",
-    "iter_schema_units",
-]
